chore: Change package name.
NOTE: we're doing this only so that we can publish a version of the
package to our GitHub Packages npm repository, as upstream are not
currently releasing this package to npmjs.org. We are not claiming
ownership of the package, and the copyright and license belong to
upstream, not Hackworth Ltd.